Add option to allow hwc_composter to use ro.sf.lcd_density as dsp DPI
This is a hack to allow hwc_composter to use ro.sf.lcd_density property
as dpi since not all kernels reports correct dpi, this allows us to
override this.

This will not give 100% correct x/y dpi, but it's all we need.

set TARGET_USES_SF_LCD_DENSITY_DPI to true to enable this.

Change-Id: I6461544042804456e57d5844caff61de4038db35